Remove VirNucPro classifier wrapper
